requirements.txt
networkx
pytest
hypothesis
python-arango
